refactor: rename section files/exports to match current naming
ArsenalSection -> SkillsSection, QuestSection -> ProjectsSection,
AccountLinksSection -> LinksSection, BackgroundOrbs -> PaperTexture.
File/export names now match what each section actually renders.
