+++ app/components/nexus/PaperTexture.tsx
